+++ src/lang/prog/ffi/js/time/mod.rs
@@ -1,0 +1,16 @@
+// devela::lang::prog::ffi::js::time
+//
+//!
+//
+
+mod instant; // JsInstant
+mod timeout; // JsTimeout
+
+crate::structural_mods! { // _mods
+    _mods {
+        pub use super::{
+            instant::*,
+            timeout::*,
+        };
+    }
+}